back/requirements.txt
djangorestframework
django
gunicorn
django-environ
environ
gevent
psycopg2
django-cors-headers